AgentGraph-RAG-SQL-Tavily/requirements.txt
Chroma
chroma-hnswlib
chromadb
gradio
ipykernel
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langsmith
numpy
openai
pandas
pydantic
pydantic_core
pypdf
pyprojroot
PyYAML
SQLAlchemy
